Refactoring and code reorganization.
