+++ quickwit-serve/src/health_check_api/handler.rs
@@ -34,7 +34,7 @@ pub enum ServiceStatus {
 }
 
 impl fmt::Display for ServiceStatus {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{:?}", self)
     }
 }
@@ -78,12 +78,16 @@ fn live_predicate(service_status: ServiceStatus) -> bool {
     matches!(service_status, ServiceStatus::Alive)
 }
 
-#[tokio::test]
-async fn test_rest_search_api_health_check_livez() {
-    let rest_search_api_filter = liveness_check_handler();
-    let resp = warp::test::request()
-        .path("/health/livez")
-        .reply(&rest_search_api_filter)
-        .await;
-    assert_eq!(resp.status(), 200);
+#[cfg(test)]
+mod tests {
+
+    #[tokio::test]
+    async fn test_rest_search_api_health_check_livez() {
+        let rest_search_api_filter = super::liveness_check_handler();
+        let resp = warp::test::request()
+            .path("/health/livez")
+            .reply(&rest_search_api_filter)
+            .await;
+        assert_eq!(resp.status(), 200);
+    }
 }
